+++ extensions/yii/jui/InputWidget.php
@@ -45,7 +45,10 @@ class InputWidget extends Widget
 	public function init()
 	{
 		if (!$this->hasModel() && $this->name === null) {
-			throw new InvalidConfigException("Either 'name' or 'model' and 'attribute' properties must be specified.");
+			throw new InvalidConfigException("Either 'name', or 'model' and 'attribute' properties must be specified.");
+		}
+		if ($this->hasModel() && !isset($this->options['id'])) {
+			$this->options['id'] = Html::getInputId($this->model, $this->attribute);
 		}
 		if($this->hasModel() && !array_key_exists('id', $this->options)) {
 			$this->options['id'] = Html::getInputId($this->model, $this->attribute);